Update install.sh
  - Customized
  - Removed Lite options since customized nanorc uses RC file wildcards in lieu of individually referenced RC files

+++ install.sh
@@ -1,53 +1,38 @@
 #!/bin/sh
 
+             ##::[[-----  Custom NanoRC Install Script -----]]::##
+
+#==============================================================================
+
 # check for unzip before we continue
 if [ ! "$(command -v unzip)" ]; then
-  echo 'unzip is required but was not found. Install unzip first and then run this script again.' >&2
+  echo '\n\n  Required, but not found: unzip \n\n  Please install unzip first, then run script again.\n\n' >&2
   exit 1
 fi
 
+# Download and extract repo:
 _fetch_sources(){
-  wget -O /tmp/nanorc.zip https://github.com/scopatz/nanorc/archive/master.zip
-  mkdir -p ~/.nano/
+  wget -O /tmp/nanorc.zip https://github.com/JW0914/nanorc-1/archive/master.zip
 
-  cd ~/.nano/ || exit
+  if [ ! -d ~/.nano ]; then
+    mkdir -p ~/.nano
+    cd ~/.nano || exit
+  fi
+
   unzip -o "/tmp/nanorc.zip"
-  mv nanorc-master/* ./
-  rm -rf nanorc-master
-  rm /tmp/nanorc.zip
+    mv nanorc-1-master/* ./
+    rm -rf nanorc-1-master
+    rm /tmp/nanorc.zip
 }
 
+# Create nanorc file:
 _update_nanorc(){
-  touch ~/.nanorc
-      
-  # add all includes from ~/.nano/nanorc if they're not already there
-  while read -r inc; do
-      if ! grep -q "$inc" "${NANORC_FILE}"; then
-          echo "$inc" >> "$NANORC_FILE"
-      fi
-  done < ~/.nano/nanorc
+  if [ ! -f ~/.nanorc ]; then
+    touch ~/.nanorc
+  fi
+
+  cat ~/.nano/nanorc >> ~/.nanorc
 }
 
-_update_nanorc_lite(){
-  sed -i '/include "\/usr\/share\/nano\/\*\.nanorc"/i include "~\/.nano\/*.nanorc"' "${NANORC_FILE}"
-}
-
-NANORC_FILE=~/.nanorc
-
-case "$1" in
- -l|--lite)
-   UPDATE_LITE=1;;
- -h|--help)
-   echo "Install script for nanorc syntax highlights"
-   echo "Call with -l or --lite to update .nanorc with secondary precedence to existing .nanorc includes"
-   exit 0
- ;;
-esac
-
-_fetch_sources;
-if [ $UPDATE_LITE ];
-then
-  _update_nanorc_lite
-else
-  _update_nanorc
-fi
+# Install:
+_fetch_sources; _update_nanorc
